+++ examples/main/src/ExampleApp.jsx
@@ -39,6 +39,7 @@ class ExampleApp extends PureComponent {
       editorResetKey: 0,
       shouldMockUpload: true,
       shouldMultiSelectImages: false,
+      shouldNativeUpload: false,
       ...localState,
     };
   }
@@ -89,7 +90,13 @@ class ExampleApp extends PureComponent {
   };
 
   initSectionsSettings = () => {
-    const { editorIsMobile, shouldMockUpload, shouldMultiSelectImages, staticToolbar } = this.state;
+    const {
+      editorIsMobile,
+      shouldMockUpload,
+      shouldMultiSelectImages,
+      staticToolbar,
+      shouldNativeUpload,
+    } = this.state;
     this.editorSettings = [
       {
         name: 'Mobile',
@@ -106,6 +113,15 @@ class ExampleApp extends PureComponent {
         action: () =>
           this.setState(state => ({
             shouldMockUpload: !state.shouldMockUpload,
+          })),
+      },
+      {
+        name: 'Native Upload Mock',
+        active: shouldNativeUpload,
+        action: () =>
+          this.setState(state => ({
+            shouldNativeUpload: !state.shouldNativeUpload,
+            editorResetKey: state.editorResetKey + 1,
           })),
       },
       {
@@ -157,6 +173,7 @@ class ExampleApp extends PureComponent {
       shouldMultiSelectImages,
       editorIsMobile,
       isToolbarShown,
+      shouldNativeUpload,
     } = this.state;
 
     return (
@@ -175,9 +192,10 @@ class ExampleApp extends PureComponent {
               <Editor
                 onChange={onEditorChange}
                 editorState={editorState}
-                isMobile={this.state.editorIsMobile || isMobile}
-                shouldMockUpload={this.state.shouldMockUpload}
-                shouldMultiSelectImages={this.state.shouldMultiSelectImages}
+                isMobile={editorIsMobile || isMobile}
+                shouldMockUpload={shouldMockUpload}
+                shouldMultiSelectImages={shouldMultiSelectImages}
+                shouldNativeUpload={shouldNativeUpload}
                 staticToolbar={staticToolbar}
                 locale={locale}
                 localeResource={localeResource}