+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 
 ### 🏠 Internal
 - `exampleApp`
+  - [#1541](https://github.com/wix-incubator/rich-content/pull/1541) file upload native\media manager toggle in gear icon
   - [#1537](https://github.com/wix-incubator/rich-content/pull/1537) bi callback `onViewerAction` outputs to console.log
 
 </details>